rename project app >>> flaskr refactor project structure as MTV block external connections to the db port 3306
