+++ docker-compose.yml
@@ -6,32 +6,39 @@ services:
       context: .
       dockerfile: Dockerfile
     volumes:
-      - ./app:/app
-    environment:
-      - FLASK_APP=app.py
+      - ./flaskr:/app
+    # environment:
+    #   - FLASK_APP=flaskr.py
     ports:
       - 3000:5000
     command: flask run --host=0.0.0.0 --reload
-    # command: python app.py
     depends_on:
-      - db
+      db:
+        condition: service_healthy
     networks:
       - my_network
 
   db:
     image: mysql:latest
-    ports:
-      - "3306:3306"
+    expose:
+      - 3306
+    # ports:
+    #   - "3306:3306"
     environment:
       - MYSQL_ROOT_PASSWORD=test
       - MYSQL_DATABASE=testdb
       - MYSQL_USER=test
       - MYSQL_PASSWORD=test
     volumes:
-      - ./app/sqldata:/var/lib/mysql
+      - ./flaskr/sqldata:/var/lib/mysql
     working_dir: /var/lib/mysql
     networks:
       - my_network
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost"]
+      interval: 20s
+      timeout: 5s
+      retries: 10
 
 networks:
   my_network:
