+++ flaskr/extensions.py
@@ -1,0 +1,5 @@
+from flask_sqlalchemy import SQLAlchemy 
+from flask_migrate import Migrate
+
+db = SQLAlchemy()
+migrate = Migrate()
